requirements.txt
PyPDF2
python-dotenv
google-genai
openai
streamlit